add exception in type and color and fix delete user

+++ src/main/java/int222/project/controllers/ColorRestController.java
@@ -12,6 +12,8 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import int222.project.exceptions.AllException;
+import int222.project.exceptions.ExceptionResponse;
 import int222.project.models.Color;
 import int222.project.repositories.ColorJpaRepository;
 
@@ -30,14 +32,26 @@ public class ColorRestController {
 		return colorJpaRepository.findById(colorId).orElse(null);
 	};
 
-	@PostMapping("/addcolor")
+	@PostMapping("/admin/addcolor")
 	public Color addColor(@RequestBody Color color) {
+		if(!colorJpaRepository.findByColorName(color.getColorName()).isEmpty()) {
+			throw new AllException(ExceptionResponse.ERROR_CODE.NAME_DUPLICATE,
+					"this name has been use"  );
+		}
 		return colorJpaRepository.save(color);
 	}
 
-	@DeleteMapping("/colordelete/{id}")
+	@DeleteMapping("/admin/colordelete/{id}")
 	public String deleteColor(@PathVariable int id) {
 		Color c=colorJpaRepository.findById(id).get();
+		if(colorJpaRepository.findById(id).isEmpty()) {
+			throw new AllException(ExceptionResponse.ERROR_CODE.NOT_NULL,
+					"this don't has this color id !"  );
+		}
+		if(!c.getProduct().isEmpty()) {
+			throw new AllException(ExceptionResponse.ERROR_CODE.NOT_NULL,
+					"this color has been use in product !"  );
+		}
 		if (c.getProduct().isEmpty()) {
 			colorJpaRepository.deleteById(id);
 			return "delete success";
@@ -45,9 +59,13 @@ public class ColorRestController {
 		else
 		return "delete fail";
 	}
-	@PutMapping("/editcolor")
+	@PutMapping("/admin/editcolor")
 	public Color editColor(@RequestBody Color color) {
 		Color old= colorJpaRepository.findById(color.getColorId()).get();
+		if(!colorJpaRepository.findByColorName(color.getColorName()).isEmpty()) {
+			throw new AllException(ExceptionResponse.ERROR_CODE.NAME_DUPLICATE,
+					"this name has been use"  );
+		}
 		old.setColorName(color.getColorName());
 		return colorJpaRepository.save(old);
 	}
